+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   the chain.
 - Support protocol version 7.
 - Support for smart contract debugging when running locally.
+- Remove JSON serialization support of BlockSummary.
+- Add an additional `indexer` to index all transaction outcomes and special events.
 
 ## 3.2.0